requirements.txt
langchain
faiss-cpu
openai
python-dotenv
langchain-community
langchain-openai
